[test] move optimizers tests to tests/it
